+++ src/main/java/com/creditscoreservice/application/creditscore/service/entityservice/CreditScoreEntityService.java
@@ -20,7 +20,7 @@ public class CreditScoreEntityService {
     }
 
 
-    public float getCreditScore() {
-        return calculateCreditScore.nextFloat();
+    public int getCreditScore() {
+        return calculateCreditScore.nextInt();
     }
 }
